+++ tests/broker/test_broker.py
@@ -176,14 +176,14 @@ async def test_invalid_login(shvbroker, url):
 
 
 async def test_invalid_hello_seq(shvbroker, url):
-    client = await RpcClientStream.connect(url.location, url.port, url.protocol)
+    client = await RpcClientStream.connect(url.location, url.port)
     await client.send(RpcMessage.request(None, "invalid"))
     with pytest.raises(RpcInvalidRequestError):
         await client.receive()
 
 
 async def test_invalid_login_seq(shvbroker, url):
-    client = await RpcClientStream.connect(url.location, url.port, url.protocol)
+    client = await RpcClientStream.connect(url.location, url.port)
     await client.send(RpcMessage.request(None, "hello"))
     await client.receive()
     await client.send(RpcMessage.request(None, "invalid"))
@@ -192,7 +192,7 @@ async def test_invalid_login_seq(shvbroker, url):
 
 
 async def test_invalid_login_null(shvbroker, url):
-    client = await RpcClientStream.connect(url.location, url.port, url.protocol)
+    client = await RpcClientStream.connect(url.location, url.port)
     await client.send(RpcMessage.request(None, "hello"))
     await client.receive()
     await client.send(RpcMessage.request(None, "login"))