+++ tests/broker/test_link_protocols.py
@@ -1,6 +1,11 @@
 """Check different transport protocols between broker and client."""
+import asyncio
 import dataclasses
+import io
 import logging
+import os
+import pty
+import sys
 
 import pytest
 
@@ -12,9 +17,13 @@ logger = logging.getLogger(__name__)
 class Protocol:
     """Generic tests based on the protocol."""
 
+    @pytest.fixture(name="broker_url")
+    async def fixture_broker_url(self, url):
+        return url
+
     @pytest.fixture(name="shvbroker")
-    async def fixture_shvbroker(self, event_loop, config, url):
-        config.listen = {"test": url}
+    async def fixture_shvbroker(self, event_loop, config, broker_url):
+        config.listen = {"test": broker_url}
         b = broker.RpcBroker(config)
         await b.start_serving()
         yield b
@@ -35,7 +44,7 @@ class TestProtocolUnix(Protocol):
 
     @pytest.fixture(name="url")
     def fixture_url(self, url, tmp_path):
-        yield dataclasses.replace(
+        return dataclasses.replace(
             url,
             protocol=RpcProtocol.LOCAL_SOCKET,
             location=str(tmp_path / "broker.sock"),
@@ -47,4 +56,4 @@ class TestProtocolUDP(Protocol):
 
     @pytest.fixture(name="url")
     def fixture_url(self, url):
-        yield dataclasses.replace(url, protocol=RpcProtocol.UDP)
+        return dataclasses.replace(url, protocol=RpcProtocol.UDP)